UPSTREAM: mb/google/brya/var/gimble{4es}: Decrease touchscreen T3 timing to 200ms
We set T3 as 300ms to meet Elan's spec, but the resume/suspend times
are greater than 500ms, which is the spec for Chromebooks.
The actual kernel timing has been measured, and given the ACPI delay
after deasserting reset in addition to the delay until the kernel
driver accesses the device, delaying only 200ms in the ACPI method is
also sufficient to meet the 300ms requirement.

BUG=b:210772498
BRANCH=none
TEST=build and test touchscreen function on DUT.
TEST=suspend, wake DUT and check touchscreen function.

Change-Id: I294296ed8e8808f3a0561ff4e95c3d5bbef59556
Signed-off-by: Nick Vaccaro <nvaccaro@google.com>
Original-Commit-Id: e24b006ee4b45fc22c0eeae618f2d32102e36148
Original-Signed-off-by: Scott Chao <scott_chao@wistron.corp-partner.google.com>
Original-Change-Id: I4bb4eda09686cb59b6e19c741aa2b78d84332d2a
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/60270
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3408228
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>