UPSTREAM: soc/intel/common/blk/memory: Make mixed topo work
When using a mixed memory topology with DDR4, it's not possible to boot
when no DIMMs are installed, even though memory-down is available. This
happens because the DIMM SPD length defaults to 256 when no DIMM SPD is
available. Relax the length check when no DIMMs are present to overcome
this problem.

Tested on system76/lemp10. Unit boots with and without DIMM installed.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Patrick Georgi <pgeorgi@google.com>
GitOrigin-RevId: ef5f7ee696097174f9d1f01e9b04ea8f1354cfe0
Change-Id: I1cabf64fade1c06a44b6c3892659d54febc7a79a
Original-Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: Tim Crawford <tcrawford@system76.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/60800
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-by: Tim Crawford <tcrawford@system76.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3379403
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>