UPSTREAM: vc/amd/fsp/sabrina: add as a copy of vc/amd/fsp/cezanne
The AMD Sabrina SoC will be using the FSP driver to call into the
corresponding FSP binary to do its part of for the silicon
initialization, so we need an initial set of FSP headers for the AMD
Sabrina SoC code to build. Since the FSP interface for this SoC won't be
too different from the Cezanne FSP interface, we'll start with a copy of
the Cezanne FSP headers and update/replace them as soon as the proper
FSP headers for Sabrina will be available.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Rob Barnes <robbarnes@google.com>
Original-Commit-Id: 6abaccf13da19733720aa424d1bd125c84ba0d85
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: Ib3bf50598efe60673b81cf99da491866fb5dc121
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/61076
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Change-Id: I1c90c5178babe00d13acbeef5e564c85348f025a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3417523
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>