UPSTREAM: libpayload: Fix legacy CBFS code after recent refactoring
The goal when adding the new CBFS API in CB:59497 was that the old CBFS
code would be left completely untouched and just moved to the side a
bit, so that it could continue to work for the payloads that use it
until they all have time to transition to the new CBFS API.
Unfortunately, between the different iterations of the patch something
went wrong with that and the final committed version of cbfs_legacy.c
does differ in some parts from the original code, including a changed
macro definition that breaks decompression support. This patch restores
all the legacy CBFS files to exactly what they used to be (other than
the necessary changes in cbfs_core.h to avoid double definition
clashes).

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Patrick Georgi <pgeorgi@google.com>
GitOrigin-RevId: db29b3765b1d760c4fff4e21b27a4aabe218f4ab
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: Ic7fd428acb03d3830f66f807cd1d7cdbd652f409
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/61061
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Jakub Czapiga <jacz@semihalf.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3389430
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>