UPSTREAM: lib/cbmem_console: Add cbmemc_copy_in function
When running in verstage before bootblock, the PSP (ARM co-processor) is
running with limited SRAM. It needs to stash the verstage console data
internally until DRAM is brought up. Once DRAM is brought up the data is
stashed in a "transfer buffer" region. In the current design, we are
using the same region for the transfer buffer and the
preram_cbmem_console region. This has the following downsides:

1) The pre-x86 buffer needs to be large enough to hold all the
   verstage, bootblock and romstage console logs.
2) On AMD platforms, the PSP verstage is signed. Changing the size of
   preram_cbmem_console after the fact will result in a mismatch of the
   transfer buffer between verstage and bootblock.

This CL adds a new method that allows SoC specific code to copy the
CBMEM console in the transfer buffer to the active CBMEM console.

BUG=b:213828947
TEST=Boot guybrush and no longer see
   *** Pre-CBMEM romstage console overflowed, log truncated!

Signed-off-by: Rob Barnes <robbarnes@google.com>
Original-Commit-Id: 1e1aa0ca4d4d5ae9d0b4917220cc72f9c441b8a2
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Original-Change-Id: Idc0ab8090db740e0d1b3d21d8968f26471f2e930
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/61099
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Kangheui Won <khwon@chromium.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Change-Id: I75d665c96a505c5a805e0dcf760f6f0b75604595
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3415829
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>