UPSTREAM: soc/intel/ehl: Add Kconfig option to disable reset on TCO expiration
The TCO timer is the default watchdog of an x86 host and can reset the
system once it has expired for the second time. There are applications
where this reset is not acceptable while the TCO is used. In these
applications the TCO expire event generates an interrupt and software
takes care. There is a bit in the TCO1_CNT register on Elkhart Lake to
prevent this reset on expiration (called NO_REBOOT, see doc #636722 ).
This bit can either be strapped on hardware or set in this register to
avoid the reset on expiration. While the hardware strap cannot be
overridden in software, the pure software solution is more flexible.
Unfortunately, the location for this bit differs among the different
platforms. This is why it has to be handled on soc level rather than on
TCO common code level.

This commit adds a Kconfig option where NO_REBOOT can be enabled. This
makes it easy to reach this feature over to the mainboard where it can
be selected if needed.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Rob Barnes <robbarnes@google.com>
Original-Commit-Id: 0099832cdab2337da5d3efc92fdd9b7a4f645cc1
Original-Change-Id: Iaa81bfbe688edd717aa02db86f0a93fecdfcd16b
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/61177
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Change-Id: If4af44a1928b48294c0109860910aa64ac21d0c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3417543
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>