UPSTREAM: soc/amd/{common,cezanne,picasso}: Add PRE_X86_CBMEM_CONSOLE_SIZE
This change splits the size of the console transfer region and size of
the bootblock/romstage Pre-RAM console region. This allows having a
larger Pre-RAM console while not impacting the size of the PSP verstage
console.

Instead of directly using the PRE_X86_CBMEM_CONSOLE_SIZE symbol in
`setup_cbmem_console`, I chose to use the offsets provided in the
transfer buffer. It would be nice to eventually do this for all the
fields in the transfer buffer.

BUG=b:213828947
TEST=Boot guybrush and verify verstage logs are no longer truncated

Signed-off-by: Rob Barnes <robbarnes@google.com>
Original-Commit-Id: 86302a806c5cc9b575424305e761753710417692
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Original-Change-Id: I8b8cc46600192a7db00f5c1f24c3c8304c4db31d
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/61189
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Kangheui Won <khwon@chromium.org>
Change-Id: I7435373d4a433468ce6eea64a525545d6f2d49d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3417510
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>