UPSTREAM: soc/intel/cannonlake: Add PcieRpHotPlug config to FSP-M
Commit b67c5ed [3rdparty/fsp: Update submodule pointer to newest master]
updated the FSP binaries/headers for Comet Lake, which included a change
moving PcieRpHotPlug from FSP-S to FSP-M. Unfortunately the existing
UDP in FSP-S was left in and deprecated, which allowed the change to go
unnoticed until it was discovered that hotplug wasn't working.

Since other related platforms (WHL, CFL) share the SoC code but use
different FSP packages, add the setting of the PcieRpHotPlug UPD to
romstage/FSP-M and guard it with '#if CONFIG(SOC_INTEL_COMETLAKE)'.

Test: build/boot Purism Librem 14, verify WiFi killswitch operates
as expected / WiFi is re-enabled when turning switch to on position.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Rob Barnes <robbarnes@google.com>
Original-Commit-Id: 74b85f2e2ec05eff9c08e6a6561210b0b00f8633
Original-Change-Id: I4e1c2ea909933ab21921e63ddeb31cefe1ceef13
Original-Signed-off-by: Matt DeVillier <matt.devillier@puri.sm>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/61377
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Michael Niewöhner <foss@mniewoehner.de>
Original-Reviewed-by: Paul Menzel <paulepanter@mailbox.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Change-Id: I035af81de234ffddbf606ae81c561e0da716abaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3424437
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Tested-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: Tim Wawrzynczak <twawrzynczak@chromium.org>