UPSTREAM: soc/amd/common/block/psp: introduce AMD_SOC_SEPARATE_EFS_SECTION
On systems that use the first 128kByte of the SPI flash for the EC
firmware, it is not possible to place the EFS/amdfw part at the lowest
location in flash where the on-chip PSP firmware will look for the EFS,
since this is at an offset of 128kByte into the flash which is where the
cbfs master header resides when the main CBFS is placed right after the
EC firmware. This patch introduces the AMD_SOC_SEPARATE_EFS_SECTION
option that allows putting the EFS in a separate FMAP section that can
be located right after the EC firmware FMAP section. The EFS FMAP
partition is checked to ensure it begins at the expected location.

(cherry picked from commit 1e25fd426ad848f79e7ee7f7de4d3dc3ca129b1f)

Signed-off-by: Fred Reitberger <reitbergerfred@gmail.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/61558
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: I5ed0f76c9c9c9c180ee5f1b96f88689d0979bb5e
GitOrigin-RevId: 1e25fd426ad848f79e7ee7f7de4d3dc3ca129b1f
Change-Id: Iec5693c100dda3a2ad9e9d60922487daa4715bbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3451344
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Andrew McRae <amcrae@google.com>
Commit-Queue: Andrew McRae <amcrae@google.com>