UPSTREAM: soc/amd/*/i2c: factor out common I2C pad configuration
The I2C pad control registers of Picasso and Cezanne are identical and
the one of Sabrina is a superset of it, so factor out the functionality.
To avoid having devicetree settings that contain raw register bits, the
i2c_pad_control struct is introduced and used. The old Picasso code for
this had the RX level hard-coded for 3.3V I2C interfaces, so keep it
this way in this patch but add a TODO  for future improvements.

(cherry picked from commit 556d1cc17f34615e3a08ccc9a48820a304a789a8)

Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/61568
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Change-Id: I1d70329644b68be3c4a1602f748e09db20cf6de1
GitOrigin-RevId: 556d1cc17f34615e3a08ccc9a48820a304a789a8
Change-Id: Idb2893be00bf32b9a3e3f27a9bf192b359ceb6e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3440305
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Andrew McRae <amcrae@google.com>
Commit-Queue: Andrew McRae <amcrae@google.com>