UPSTREAM: nb,soc/amd/*/iommu: fix comment about IOMMU MMIO resource
This comment was added with the AMD family 15h Trinity IOMMU support in
commit 88ebbeb7e2a914330c869147bacb190b4270532f and looks like a copy of
the comment about the subtractive decode ranges in the LPC device. The
IOMMU doesn't have any subtractively decoded I/O or MMIO ranges and this
is also not what the code does. This resource is the MMIO region to
configure the IOMMU instead, so fix the comment in all copies of the
IOMMU support code.

(cherry picked from commit ddf137f82261bfcbf394daf413dd8da920786c0d)

Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/61631
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Change-Id: I2e1e3a46b839b9e58b836932c1bc9b41b1b1dc02
GitOrigin-RevId: ddf137f82261bfcbf394daf413dd8da920786c0d
Change-Id: I11b5cb7602621a261a8d81fe56825593f5e1df55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3442634
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Andrew McRae <amcrae@google.com>
Commit-Queue: Andrew McRae <amcrae@google.com>