UPSTREAM: payloads/tianocore: Pass SD_MMC_TIMEOUT build option
By default, edk2 allows 1000000μs for SD Card Readers and eMMC to
initialize which is excessive and causes a boot delay. This makes
the value configurable and uses a default of 1000μs which is sufficient
for the majority of readers. The value of 1000μs was hardcoded in
MrChromeBox's fork for around 2 years with no reported issues.

(cherry picked from commit 4a9be9f32120794e7ea16d7dd5f4490f17dad623)

Signed-off-by: Sean Rhodes <sean@starlabs.systems>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/61902
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Change-Id: I873bcddf6f37a9eaae5c84991b3996d51fb460d3
GitOrigin-RevId: 4a9be9f32120794e7ea16d7dd5f4490f17dad623
Change-Id: I3cd6182b541a1d289a42388092f7d92cb07323fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3492037
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>