UPSTREAM: drivers/intel/i210: Set log level to BIOS_NOTICE on missing MAC address
Set the log level to BIOS_NOTICE for the case where the mainboard can
not provide a MAC address since this can be a valid case. Showing this
message with log level BIOS_ERR is not appropriate.

In addition, rephrase the message to make clear that if the mainboard
does not provide a MAC address the one stored in the MAC will be used.

(cherry picked from commit 4b2490eed629801cfb59dc2bf20cccebba8daa0a)

Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/62008
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Mario Scheithauer <mario.scheithauer@siemens.com>
Original-Change-Id: Ibfc58845f0ea47ced048b446e685c4860a29f075
GitOrigin-RevId: 4b2490eed629801cfb59dc2bf20cccebba8daa0a
Change-Id: I062a654a37fc702a0635e6e755163d356af0df2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3513113
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Rob Barnes <robbarnes@google.com>